third_party/WebKit/LayoutTests/imported/web-platform-tests/html/dom/documents/dom-tree-accessors/document.forms-expected.txt

This is a testharness.js-based test.
FAIL document.forms assert_equals: expected "form2" but got "form1"
PASS document.forms with empty string 
FAIL document.forms iteration assert_array_equals: lengths differ, expected 6 got 8
PASS document.forms getOwnPropertyNames 
Harness: the test ran to completion.
